Correct for synonym values
Usually in the intent_info, the JSON rawValue and value are the same for entity name and state. However, if synonyms are used in names or states, rawValue will not be correct for Hubitat control. For example, for shade control 'lower' is a synonym for 'close', so we need to use the 'close' command for the shade. This corrects deriving the name and state.